PowerMenu: Only tint items once
Google and in their infinite wisdom decided that they would tint
power menu items twice. The problem with this is that in our case
with the 'advanced' reboot item, it would case the item to be darker
than the rest. Not to mention that is completely unnecessary to do
this, just wasteful lol

- Also fix the 'screen record' item so that it tints correctly. We
didn't catch this because Google's tint was a fallback :p

Test: Enable all power menu items and press and hold power button.
Cycle through light/dark theme.

Change-Id: I0a6d1543e9591181c15f1d0acff87a818a67fa1b
Signed-off-by: Ayush Dubey <ayushdubey70@gmail.com>